--- lib/aws/session_store/dynamo_db/errors/missing_secret_key_error.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-
-module Aws::SessionStore::DynamoDB::Errors
-  # This error is raised when no secret key is provided.
-  class MissingSecretKeyError < RuntimeError
-    def initialize(msg = 'No secret key provided!')
-      super
-    end
-  end
-end
